Split debugger_cpu and device_debug into separate sources (nw)
