+++ src/devices/cpu/i386/i386.cpp
@@ -22,7 +22,7 @@
 #include "i386priv.h"
 #include "i386.h"
 
-#include "debug/debugcpu.h"
+#include "debug/devdebug.h"
 
 /* seems to be defined on mingw-gcc */
 #undef i386
