+++ src/emu/debug/dvwpoints.h
@@ -12,7 +12,7 @@
 #define __DVWPOINTS_H__
 
 #include "debugvw.h"
-#include "debugcpu.h"
+#include "devdebug.h"
 
 
 //**************************************************************************
