+++ src/osd/modules/debugger/osx/disassemblyviewer.mm
@@ -15,6 +15,7 @@
 #include "debugger.h"
 #include "debug/debugcon.h"
 #include "debug/debugcpu.h"
+#include "debug/devdebug.h"
 
 
 @implementation MAMEDisassemblyViewer
