+++ src/osd/modules/debugger/qt/mainwindow.cpp
@@ -12,6 +12,7 @@
 
 #include "debug/debugcon.h"
 #include "debug/debugcpu.h"
+#include "debug/devdebug.h"
 #include "debug/dvdisasm.h"
