+++ src/osd/modules/debugger/win/consolewininfo.cpp
@@ -14,6 +14,7 @@
 #include "debugger.h"
 #include "debug/debugcon.h"
 #include "debug/debugcpu.h"
+#include "debug/devdebug.h"
 #include "imagedev/cassette.h"
 
 #include "strconv.h"
