+++ src/osd/modules/debugger/win/disasmbasewininfo.cpp
@@ -14,6 +14,7 @@
 #include "debugger.h"
 #include "debug/debugcon.h"
 #include "debug/debugcpu.h"
+#include "debug/devdebug.h"
 
 //#include "winutf8.h"
